requirements.txt
httpx
Pillow
